Updated DataReader
- DataReader now uses Precision instead of short int
- DataReader collects chunks of data before writing to the buffer
- DataReader is created in the main function and should load data

+++ src/Types.h
@@ -6,7 +6,7 @@
 #include "Constants.h"
 #include "Ringbuffer.h"
 
-typedef short int Precision;
+typedef float Precision;
 typedef Precision Sample[SAMPLE_COUNT];
 typedef Sample SampleChunk[CHUNK_COUNT];
 struct fitData {
